refactor(wizard-eda): created tabs in EDA page, moved plot related stuff into own folder, renamed explore to EDA
