+++ eis_qgis_plugin/resources/ui/eda/wizard_explore_old.ui
