+++ eis_qgis_plugin/resources/ui/eda/wizard_plot_ecdf.ui
