+++ eis_qgis_plugin/wizard/eda/plots/lineplot.py
@@ -8,9 +8,9 @@ from qgis.PyQt.QtWidgets import QWidget
 
 import eis_qgis_plugin.libs.seaborn as sns
 from eis_qgis_plugin.qgis_plugin_tools.tools.resources import load_ui
-from eis_qgis_plugin.wizard.plots.plot_template import EISPlot
+from eis_qgis_plugin.wizard.eda.plots.plot_template import EISPlot
 
-FORM_CLASS: QWidget = load_ui("explore/wizard_plot_lineplot.ui")
+FORM_CLASS: QWidget = load_ui("eda/wizard_plot_lineplot.ui")
 
 
 class EISWizardLineplot(EISPlot, FORM_CLASS):
