+++ eis_qgis_plugin/wizard/eda/plots/plot_template.py
