+++ eis_qgis_plugin/wizard/eda/plots/scatterplot.py
@@ -1,36 +1,44 @@
-from qgis.core import QgsMapLayerProxyModel
-from qgis.gui import QgsColorButton, QgsFieldComboBox
-from qgis.PyQt.QtWidgets import QComboBox, QWidget
+from qgis.core import QgsFieldProxyModel, QgsMapLayerProxyModel
+from qgis.gui import (
+    QgsColorButton,
+    QgsFieldComboBox,
+    QgsOpacityWidget,
+)
+from qgis.PyQt.QtWidgets import QWidget
 
 import eis_qgis_plugin.libs.seaborn as sns
 from eis_qgis_plugin.qgis_plugin_tools.tools.resources import load_ui
-from eis_qgis_plugin.wizard.plots.plot_template import EISPlot
+from eis_qgis_plugin.wizard.eda.plots.plot_template import EISPlot
 
-FORM_CLASS: QWidget = load_ui("explore/wizard_plot_boxplot.ui")
+FORM_CLASS: QWidget = load_ui("eda/wizard_plot_scatterplot.ui")
 
 
-class EISWizardBoxplot(EISPlot, FORM_CLASS):
+class EISWizardScatterplot(EISPlot, FORM_CLASS):
     """
-    Class for EIS-Seaborn boxplots.
+    Class for EIS-Seaborn scatterplots.
 
     Initialized from a UI file. Responsible for updating widgets and
     producing the plot.
     """
 
     def __init__(self, parent=None) -> None:
-        
+
         # DECLARE TYPES
         self.X: QgsFieldComboBox
         self.Y: QgsFieldComboBox
 
         self.color_field: QgsFieldComboBox
         self.color: QgsColorButton
-        self.log_scale: QComboBox
+        self.opacity: QgsOpacityWidget
+        self.size: QgsFieldComboBox
+        self.style: QgsFieldComboBox
 
         # Initialize
         self.collapsed_height = 200
         super().__init__(parent)
         self.layer.setFilters(QgsMapLayerProxyModel.VectorLayer)
+        self.X.setFilters(QgsFieldProxyModel.Filter.Numeric)
+        self.Y.setFilters(QgsFieldProxyModel.Filter.Numeric)
 
         self.update_layer(self.layer.currentLayer())
 
@@ -43,6 +51,8 @@ class EISWizardBoxplot(EISPlot, FORM_CLASS):
         self.X.setLayer(layer)
         self.Y.setLayer(layer)
         self.color_field.setLayer(layer)
+        self.size.setLayer(layer)
+        self.style.setLayer(layer)
 
 
     def plot(self, ax):
@@ -52,21 +62,36 @@ class EISWizardBoxplot(EISPlot, FORM_CLASS):
         X_field_name = self.X.currentField()
         Y_field_name = self.Y.currentField()
         color_field_name = self.color_field.currentField()
+        size_field_name = self.size.currentField()
+        style_field_name = self.style.currentField()
         fields = [X_field_name, Y_field_name]
         if color_field_name:
             fields.append(color_field_name)
+        if size_field_name:
+            fields.append(size_field_name)
+        if style_field_name:
+            fields.append(style_field_name)
 
         df = self.vector_layer_to_df(layer, *fields)
 
         # if color_field_name:
         #     self.check_unique_values(df, color_field_name, 20)
 
-        sns.boxplot(
+        if size_field_name:
+            self.check_unique_values(df, size_field_name, 10)
+
+        if style_field_name:
+            self.check_unique_values(df, style_field_name, 10)
+
+        sns.scatterplot(
             data=df,
             x=X_field_name,
             y=Y_field_name,
             hue=color_field_name if color_field_name else None,
             color=self.color.color().getRgbF(),
+            alpha=self.opacity.opacity(),
+            size=size_field_name if size_field_name else None,
+            style=style_field_name if style_field_name else None,
             ax=ax
         )
 
@@ -75,10 +100,10 @@ class EISWizardBoxplot(EISPlot, FORM_CLASS):
         """Produce example plot using SNS data."""
         penguins = sns.load_dataset("penguins")
 
-        sns.boxplot(
+        sns.scatterplot(
             data=penguins,
             x="flipper_length_mm",
-            y="species",
+            y="bill_length_mm",
             hue="species",
             palette="dark",
             ax=ax
@@ -90,4 +115,6 @@ class EISWizardBoxplot(EISPlot, FORM_CLASS):
         super().reset()
 
         self.color_field.setField("")
-        self.log_scale.setCurrentIndex(0)
+        self.opacity.setOpacity(100)
+        self.size.setField("")
+        self.style.setField("")
